# internal/feature/tasks/transport/http/dto_common.go
package task_transport_http

import (
	"time"

	"github.com/FiL4an/golang-todoapp/internal/core/domain"
)

type TaskDTOResponse struct {
	ID           int        `json:"id" example:"4"`
	Version      int        `json:"version" example:"3"`
	Title        string     `json:"title" example:"Дошака"`
	Description  *string    `json:"description" example:"Сделать до четверга домашнее задание по математике"`
	Completed    bool       `json:"completed" example:"false"`
	CreateAt     time.Time  `json:"created_at" example:"2026-02-26T10:30:00Z"`
	CompletedAt  *time.Time `json:"completed_at" example:"null"`
	AuthorUserID int        `json:"author_user_id" example:"5"`
}

func taskDTOFromDomain(task domain.Task) TaskDTOResponse {
	return TaskDTOResponse{
		ID:           task.ID,
		Version:      task.Version,
		Title:        task.Title,
		Description:  task.Description,
		Completed:    task.Completed,
		CreateAt:     task.CreatedAt,
		CompletedAt:  task.CompletedAt,
		AuthorUserID: task.AuthorUserID,
	}
}
func taskDTOsFromDomains(tasks []domain.Task) []TaskDTOResponse {
	dtos := make([]TaskDTOResponse, len(tasks))

	for i, task := range tasks {
		dtos[i] = taskDTOFromDomain(task)
	}
	return dtos
}